+++ GhostHunter/src/uva/cs2110/ghosthunter/Projectile.java
@@ -2,9 +2,9 @@ package uva.cs2110.ghosthunter;
 
 public interface Projectile {
 	
-	public float getX();
+	public int getX();
 	
-	public float getY();
+	public int getY();
 	
 	public void updateX();
